Remove scripting and activeTab permission.
Version 0.1.0 was rejected by Chrome Web Store due to "scripting"
permission was requested but not used.  I asked ChatGPT 5 and it said it
was not used.  Why did it write it in the first place...

It also said "activeTab" can be removed.  I tested it and it seemed to
work fine.